+++ src/entities/projectiles/venom_spit.py
@@ -11,7 +11,7 @@ class VenomSpit(BallisticProjectile):
     rect: pygame.Rect | pygame.FRect
 
     def __init__(
-        self, damage: int, x_position: float, y_position: float, target: Enemy
+        self, damage: float, x_position: float, y_position: float, target: Enemy
     ) -> None:
         image = load_asset("venom_spit")
